# packages/contracts/openapi/core-api.v1.yaml
openapi: 3.1.0
...
info:
  title: AuthWeave Core API
  version: 1.0.0
  description: >-
    Workspace-scoped assessment boundary owned by the deterministic AuthWeave core.
    The local Phase 2 API is intentionally unauthenticated and refuses to start with a
    non-loopback server address. Workspace provisioning and every assessment operation
    require authenticated authorization before a non-local deployment. Do not expose
    this local service through a proxy or tunnel.
servers:
  - url: http://127.0.0.1:8080
    description: Local Core API
security: []
...
paths:
  /api/v1/workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    put:
      tags:
        - Workspaces
      summary: Provision a workspace
      description: >-
        Idempotently provisions a workspace for the local Phase 2 flow. Authentication
        provisioning will own this action in a later phase.
      operationId: provisionWorkspaceV1
      responses:
        "201":
          description: Workspace created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "204":
          description: Workspace already exists.
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /api/v1/workspaces/{workspaceId}/assessments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags:
        - Assessments
      summary: Create a draft assessment
      operationId: createAssessmentV1
      responses:
        "201":
          description: Draft assessment created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v1.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
  /api/v1/workspaces/{workspaceId}/assessments/{assessmentId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    get:
      tags:
        - Assessments
      summary: Read an assessment
      operationId: getAssessmentV1
      responses:
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
        "200":
          description: Current assessment representation.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v1.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
  /api/v1/workspaces/{workspaceId}/assessments/{assessmentId}/profile:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    put:
      tags:
        - Assessments
      summary: Replace an assessment profile
      description: >-
        Replaces the complete structured profile only when expectedVersion matches the
        persisted assessment version. An unchanged profile preserves status and version;
        a stale expectedVersion still conflicts. Archived assessments cannot be changed.
        A successful change atomically appends a revision and a state-change event.
        Unchanged or rejected requests append neither.
      operationId: updateAssessmentProfileV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/update-assessment-profile-request.v1.schema.json
      responses:
        "200":
          description: Current profile; version increments only when the profile changes.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v1.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "422":
          $ref: "#/components/responses/InvalidProfile"
  /api/v1/workspaces/{workspaceId}/assessments/{assessmentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
      - $ref: "#/components/parameters/AfterVersion"
      - $ref: "#/components/parameters/HistoryLimit"
    get:
      tags: [Assessments]
      summary: Read immutable assessment revisions
      description: >-
        Returns snapshots in ascending version order. Existing assessments begin with
        a BASELINE at their pre-migration version; earlier states are not reconstructed.
        Pagination is forward-only, not a frozen snapshot across requests. This endpoint
        shares the local-only restriction of the rest of this API.
      operationId: getAssessmentRevisionsV1
      responses:
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
        "200":
          description: Revision page; nextAfterVersion is null when no more rows were found.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-revision-page.v1.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
...
  /api/v2/workspaces/{workspaceId}/assessments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Assessments]
      summary: Create a draft with a profile v2 representation
      description: >-
        Provision the workspace through the existing v1 workspace endpoint first.
        Empty residency details mean unrecorded. The database uses the smallest lossless
        profile format; the response always projects profile v2. This API is local-only.
      operationId: createAssessmentV2
      responses:
        "201":
          description: Created assessment with empty residency details.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v2.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
  /api/v2/workspaces/{workspaceId}/assessments/{assessmentId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    get:
      tags: [Assessments]
      summary: Read a current assessment as profile v2
      description: >-
        Projects legacy profiles with empty residency details without updating stored
        data, timestamps, version or history. profileSchemaVersion describes the response
        representation. The original criticality is preserved. Existing v1 reads return
        409 profile-upgrade-required when residency details cannot be represented.
        V2 reads return the same code when recorded v3 authentication controls require v3.
      operationId: getAssessmentV2
      responses:
        "200":
          description: Current assessment with an explicit v2 profile.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v2.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
  /api/v2/workspaces/{workspaceId}/assessments/{assessmentId}/profile:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    put:
      tags: [Assessments]
      summary: Replace a profile including scoped country residency inputs
      description: >-
        Requires the complete profile and expectedVersion. Countries are an allowlist
        for at-rest storage of every selected data category. Empty lists are unrecorded,
        not unrestricted. Processing locations, remote access, cross-border transfers
        and compliance are not represented by this input. REQUIRED and PREFERRED retain
        their existing meaning; FORBIDDEN is not interpreted as a country denylist.
        Scoped residency evaluation is available through the v2 eligibility-preflight
        endpoint; v1 preflights continue to defer it.
        Reads/no-op saves do not rewrite old snapshots. Changes atomically append a
        versioned snapshot and security-section event. V1 writes return 409 when the
        current profile contains residency details, preventing accidental data loss.
        V2 writes likewise reject recorded authentication controls; use v3 to preserve them.
      operationId: updateAssessmentProfileV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/update-assessment-profile-request.v2.schema.json
      responses:
        "200":
          description: Updated or unchanged assessment.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v2.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "422":
          $ref: "#/components/responses/InvalidProfile"
  /api/v2/workspaces/{workspaceId}/assessments/{assessmentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
      - $ref: "#/components/parameters/AfterVersion"
      - $ref: "#/components/parameters/HistoryLimit"
    get:
      tags: [Assessments]
      summary: Read original v1 and v2 profile snapshots
      description: >-
        Preserves each stored profile and its profileSchemaVersion without applying
        current domain validation or projecting old snapshots into a newer shape.
        Uses the same exclusive version cursor as v1. A v1 history request returns 409
        if its page contains a v2/v3 snapshot; v2 history returns 409 for pages containing v3.
        State-change events keep their existing v1 endpoint.
      operationId: getAssessmentRevisionsV2
      responses:
        "200":
          description: A page containing original versioned snapshots.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-revision-page.v2.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
...
  /api/v3/workspaces/{workspaceId}/assessments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Assessments]
      summary: Create a draft with a profile v3 representation
      description: >-
        Provision the workspace through the existing v1 endpoint first. Projects empty residency details and UNKNOWN authentication controls. The stored profile uses the smallest lossless format; this local-only response always projects v3.
      operationId: createAssessmentV3
      responses:
        "201":
          description: Created assessment with unrecorded residency and authentication controls.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v3.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
  /api/v3/workspaces/{workspaceId}/assessments/{assessmentId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    get:
      tags: [Assessments]
      summary: Read a current assessment as profile v3
      description: >-
        Projects v1/v2 profiles without rewriting data, timestamps, versions or history. UNKNOWN controls do not follow from assurance labels. profileSchemaVersion identifies the response representation. Older current-profile endpoints return 409 if they cannot preserve recorded requirements. This v3 endpoint also returns 409 for a recorded v4 compliance scope.
      operationId: getAssessmentV3
      responses:
        "200":
          description: Current assessment with an explicit v3 profile.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v3.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
  /api/v3/workspaces/{workspaceId}/assessments/{assessmentId}/profile:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    put:
      tags: [Assessments]
      summary: Replace a profile including independent authentication controls
      description: >-
        Replace the complete profile with expectedVersion. Authentication controls apply independently to all selected human clients and populations; no automatic assurance-level or NIST AAL mapping. REQUIRED needs enforceability, not mere availability. Step-up means stronger authentication, not repeating the same method. PREFERRED is not yet scored; FORBIDDEN needs intent clarification. Residency rules remain unchanged. No-op saves preserve storage; real changes append an atomic snapshot and security-section event. Older v1/v2 writes cannot silently erase recorded controls. V3 writes reject recorded v4 compliance scope.
      operationId: updateAssessmentProfileV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/update-assessment-profile-request.v3.schema.json
      responses:
        "200":
          description: Updated or unchanged assessment.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v3.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "422":
          $ref: "#/components/responses/InvalidProfile"
  /api/v3/workspaces/{workspaceId}/assessments/{assessmentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
      - $ref: "#/components/parameters/AfterVersion"
      - $ref: "#/components/parameters/HistoryLimit"
    get:
      tags: [Assessments]
      summary: Read original v1, v2 and v3 profile snapshots
      description: >-
        Return exact mixed v1/v2/v3 snapshots and original schema versions, with the existing exclusive version cursor. No history upcasting or rewriting. Older history endpoints reject a page containing a newer unsupported format, including v4 scope snapshots. Events keep their v1 endpoint.
      operationId: getAssessmentRevisionsV3
      responses:
        "200":
          description: A page containing original versioned snapshots.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-revision-page.v3.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
...
  /api/v4/workspaces/{workspaceId}/assessments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Assessments]
      summary: Create a draft with a profile v4 representation
      description: >-
        Provision the workspace through the existing v1 endpoint first. A new draft has UNKNOWN complianceScopeStatus, empty target labels, empty residency details and UNKNOWN authentication controls. This local-only response projects profile v4; storage uses the smallest lossless format.
      operationId: createAssessmentV4
      responses:
        "201":
          description: Created assessment with unrecorded requirements scope.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v4.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
  /api/v4/workspaces/{workspaceId}/assessments/{assessmentId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    get:
      tags: [Assessments]
      summary: Read a current assessment as profile v4
      description: >-
        Projects legacy profiles with UNKNOWN complianceScopeStatus without changing targets, timestamps, versions or history. Empty legacy target lists do not imply NONE_IDENTIFIED. Existing target labels do not confirm scope or provider compliance. The response profileSchemaVersion describes the representation. Recorded v5 usage inputs require the v5 endpoint.
      operationId: getAssessmentV4
      responses:
        "200":
          description: Current assessment with an explicit v4 profile.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v4.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
  /api/v4/workspaces/{workspaceId}/assessments/{assessmentId}/profile:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    put:
      tags: [Assessments]
      summary: Replace a profile including an explicit compliance requirements scope
      description: >-
        Replace the complete profile with expectedVersion. UNKNOWN preserves an unrecorded or incomplete requirements scope, including existing target lists. NONE_IDENTIFIED requires an empty target list; TARGETS_IDENTIFIED requires at least one label. Contradictions return 422, malformed fields return 400. These are owner-recorded inputs, not legal applicability or compliance findings. No-op saves do not rewrite storage. Real changes append an atomic security-section event and snapshot. V4 writes reject recorded v5 usage inputs. V1/v2/v3 writes cannot silently clear recorded scope; an intentional v4 clear to UNKNOWN can restore older current-profile compatibility without altering history.
      operationId: updateAssessmentProfileV4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/update-assessment-profile-request.v4.schema.json
      responses:
        "200":
          description: Updated or unchanged assessment.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v4.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "422":
          $ref: "#/components/responses/InvalidProfile"
  /api/v4/workspaces/{workspaceId}/assessments/{assessmentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
      - $ref: "#/components/parameters/AfterVersion"
      - $ref: "#/components/parameters/HistoryLimit"
    get:
      tags: [Assessments]
      summary: Read original v1, v2, v3 and v4 profile snapshots
      description: >-
        Return exact mixed v1/v2/v3/v4 snapshots and their original schema versions, using the existing exclusive version cursor. No history projection, rewriting or current-domain revalidation. Older APIs reject pages containing unsupported formats. Events remain on the v1 endpoint.
      operationId: getAssessmentRevisionsV4
      responses:
        "200":
          description: A page containing original versioned snapshots.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-revision-page.v4.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/ProfileUpgradeRequired"
...
  /api/v5/workspaces/{workspaceId}/assessments:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Assessments]
      summary: Create a draft with a profile v5 representation
      description: >-
        Provision the workspace through the existing v1 endpoint. A new draft has empty usagePlanning inputs; no traffic values or costs are inferred. This response projects profile v5 while storage uses the smallest lossless format. This API is local-only.
      operationId: createAssessmentV5
      responses:
        "201":
          description: Created assessment with unrecorded usage planning.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v5.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/WorkspaceNotFound"
  /api/v5/workspaces/{workspaceId}/assessments/{assessmentId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    get:
      tags: [Assessments]
      summary: Read a current assessment as profile v5
      description: >-
        Projects older profiles with an empty usagePlanning object without changing records, timestamps, lock versions or history. Empty metric maps mean unknown, not zero. The response profileSchemaVersion describes this representation, not a storage migration.
      operationId: getAssessmentV5
      responses:
        "200":
          description: Current assessment with an explicit v5 profile.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v5.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
  /api/v5/workspaces/{workspaceId}/assessments/{assessmentId}/profile:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
    put:
      tags: [Assessments]
      summary: Replace a profile including usage volumes and planning assumptions
      description: >-
        Replace the complete profile with expectedVersion. Usage counts require an explicit ASSUMED or OBSERVED basis and a non-negative JavaScript-safe integer. Missing metric keys remain unknown; zero must be supplied explicitly. Scope and assumptions are bounded owner-supplied text, never executable instructions. Incomplete planning inputs are allowed in drafts. No-op saves preserve data; changes append an atomic operations-section event and versioned snapshot. Older API writes cannot silently erase recorded planning inputs. Explicit v5 clearing restores the smallest lossless stored format without changing older history. No cost estimate or provider ranking is produced.
      operationId: updateAssessmentProfileV5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/update-assessment-profile-request.v5.schema.json
      responses:
        "200":
          description: Updated or unchanged assessment.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-response.v5.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
        "422":
          $ref: "#/components/responses/InvalidProfile"
  /api/v5/workspaces/{workspaceId}/assessments/{assessmentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/AssessmentId"
      - $ref: "#/components/parameters/AfterVersion"
      - $ref: "#/components/parameters/HistoryLimit"
    get:
      tags: [Assessments]
      summary: Read original v1 through v5 profile snapshots
      description: >-
        Return exact mixed v1/v2/v3/v4/v5 snapshots with their original schema versions. No history projection, rewriting or current-domain validation. Older history APIs reject pages containing unsupported formats; events keep the v1 endpoint.
      operationId: getAssessmentRevisionsV5
      responses:
        "200":
          description: A page containing original versioned snapshots.
          content:
            application/json:
              schema:
                $ref: ../schemas/assessment-revision-page.v5.schema.json
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/AssessmentNotFound"
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace boundary for every assessment query.
      schema:
        type: string
        format: uuid
    AssessmentId:
      name: assessmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AfterVersion:
      name: afterVersion
      in: query
      description: Exclusive version cursor. Omit to start with the earliest stored version.
      schema:
        type: integer
        minimum: 0
        maximum: 9007199254740991
    HistoryLimit:
      name: limit
      in: query
      description: Maximum number of history rows returned per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  headers:
    Location:
      description: Relative URI of the created resource.
      schema:
        type: string
        format: uri-reference
  responses:
    ProfileUpgradeRequired:
      description: This representation cannot preserve newer requirements; use the compatible v2, v3, v4 or v5 endpoint.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json
    InvalidRequest:
      description: The request body, path or query parameters are malformed or outside the documented bounds.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json
    WorkspaceNotFound:
      description: The workspace has not been provisioned.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json
    AssessmentNotFound:
      description: The assessment does not exist inside the supplied workspace.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json
    VersionConflict:
      description: The expected version is stale, the assessment is archived, or an older client must use a newer profile API.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json
    InvalidProfile:
      description: The profile contains one or more domain contradictions.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/core-problem.v1.schema.json